// source/frontend/src/models/api/generated/ApiGen_Concepts_DispositionFileTeam.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Concepts_FileTeam } from './ApiGen_Concepts_FileTeam';

// LINK: @backend/apimodels/Models/Concepts/DispositionFile/DispositionFileTeamModel.cs
export interface ApiGen_Concepts_DispositionFileTeam extends ApiGen_Concepts_FileTeam {
  dispositionFileId: number;
  parentFileId: number;
}